Split header and nav components
