+++ app/javascript/app/components/header/header.js
@@ -1,0 +1,9 @@
+import { connect } from 'react-redux';
+
+import Component from './header-component';
+
+const mapStateToProps = ({ location }) => ({
+  routes: Object.values(location.routesMap).filter(r => !!r.nav)
+});
+
+export default connect(mapStateToProps, null)(Component);
