+++ app/javascript/app/components/nav/nav-styles.scss
@@ -1,73 +1,4 @@
 @import '~styles/settings.scss';
-@import '~styles/layout.scss';
-
-$logo-width: 215px;
-$flag-height: 40px;
-
-// Header wrapper
-.headerContainer {
-  background-color: $white;
-}
-
-.header {
-  padding-top: 20px;
-
-  @include row();
-}
-
-.navbarContainer {
-  display: flex;
-  align-items: center;
-}
-
-// True for every side of header
-
-.navElement {
-  flex: 1;
-  display: flex;
-  align-items: center;
-}
-
-// Left side of the header
-
-.logo {
-  width: $logo-width;
-}
-
-// Center of the header
-
-.pageTitleContainer {
-  font-size: $font-size;
-  font-family: $font-family-2;
-  text-transform: uppercase;
-  letter-spacing: 2px;
-  justify-content: center;
-  flex-grow: 3;
-
-  @media #{$tablet-landscape} {
-    font-size: 30px;
-  }
-}
-
-.climatText {
-  color: $camarone;
-}
-
-.reportText {
-  color: $dusty-gray;
-}
-
-// Right side of the header
-
-.flagContainer {
-  justify-content: flex-end;
-}
-
-.flag {
-  height: $flag-height;
-}
-
-// Link styles
 
 .link {
   color: $theme-color;
@@ -107,90 +38,5 @@ $flag-height: 40px;
     position: absolute;
     bottom: 0;
     left: 0;
-  }
-}
-
-// Tabs Container
-
-.tabsContainer {
-  display: flex;
-  align-items: center;
-  margin: {
-    top: 20px;
-  }
-
-  @media #{$tablet-landscape} {
-    justify-content: center;
-  }
-}
-
-// Sticky navbar
-
-.stickyNavElement {
-  display: flex;
-}
-
-.smallLogo {
-  display: none;
-  height: 35px;
-}
-
-.smallFlag {
-  display: none;
-  height: 25px;
-}
-
-// Sticky navbar styles when it's visible
-.stickyWrapper {
-  max-height: 55px;
-
-  .header {
-    padding-top: 0;
-  }
-
-  .navbarContainer {
-    display: none;
-  }
-
-  .smallLogo {
-    display: flex;
-  }
-
-  .smallFlag {
-    display: flex;
-  }
-
-  .stickyNavElement {
-    flex: 1;
-    display: flex;
-  }
-
-  .stickyImageElement {
-    height: 100%;
-    align-items: center;
-  }
-
-  .stickyTabs {
-    flex: auto;
-    flex-grow: 1;
-    justify-content: center;
-    height: 100%;
-  }
-
-  .stickyFlag {
-    justify-content: flex-end;
-  }
-
-  .tabsContainer {
-    height: 55px;
-    margin-top: 0;
-  }
-
-  .link {
-    padding-bottom: 0;
-
-    &.active::after {
-      bottom: 0;
-    }
   }
 }
